refactor: practice folder -> typing (naming fix)
