+++ client/src/components/typing/long/Layout/index.tsx
