+++ client/src/components/typing/long/Typing/InfoBar/InfoBarItem.tsx
