+++ client/src/components/typing/long/Typing/TypingHeader/index.tsx
