+++ client/src/components/typing/long/Typing/index.tsx
@@ -7,10 +7,10 @@ import { useRecoilValue } from 'recoil';
 import { getTypingHistoryAPI } from '@/apis/typing';
 import { userProfileState } from '@/atom/userProfile';
 import PracticeResultModal from '@/components/common/ResultModal/practice-mode';
-import PracticeLongLayout from '@/components/practice/long/Layout';
-import TypingHeader from '@/components/practice/long/Typing/TypingHeader/index';
-import TypingDetailPagination from '@/components/practice/long/TypingDetailPagination';
-import TypingLine from '@/components/practice/long/TypingLine';
+import PracticeLongLayout from '@/components/typing/long/Layout';
+import TypingHeader from '@/components/typing/long/Typing/TypingHeader/index';
+import TypingDetailPagination from '@/components/typing/long/TypingDetailPagination';
+import TypingLine from '@/components/typing/long/TypingLine';
 import { PRACTICE_LONG_PATH, PRACTICE_LONG_PATH_DETAIL } from '@/constants/paths';
 import useStopwatch from '@/hooks/useStopWatch';
 import useToggle from '@/hooks/useToggle';
