+++ client/src/components/typing/long/TypingChar/index.tsx
