+++ client/src/components/typing/long/TypingDetailPagination/index.stories.tsx
