+++ client/src/components/typing/long/TypingDetailPagination/index.tsx
