+++ client/src/components/typing/long/TypingExit/index.stories.tsx
