+++ client/src/components/typing/long/TypingExit/index.tsx
