+++ client/src/components/typing/short/InfoBar/GrassEllipse.tsx
