+++ client/src/components/typing/short/InfoBar/InfoBarItem.tsx
