+++ client/src/components/typing/short/OriginalTyping/index.tsx
