+++ client/src/components/typing/short/_hook/useCurrentTyping.ts
